Fixed production issues

+++ server/src/seeds/cleanDb.ts
@@ -3,9 +3,12 @@ import db from '../config/connection.js';
 
 export default async (modelName: "Question", collectionName: string) => {
   try {
+    if (!models[modelName] || !models[modelName].db) {
+      throw new Error(`Model or database not found for modelName: ${modelName}`);
+    }
     let modelExists = await models[modelName].db.db.listCollections({
       name: collectionName
-    }).toArray()
+    }).toArray();
 
     if (modelExists.length) {
       await db.dropCollection(collectionName);
